KVM: SVM: Provide some SEV-ES debugging capabilities
Debugging patch to dump a GHCB or a VMCB/VMSA of an SEV-ES guest. The
debug policy bit must be enabled for the guest in order to successfully
decrypt the VMSA.

Also, provide a module parameter (that can be set dynamically through
sysfs - defaults to off) that can be used to dump all vCPU VMCB/VMSA areas
before VMSA encryption and on guest termination.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>