KVM: SEV: Add KVM_SEV_SNP_LAUNCH_FINISH command
The KVM_SEV_SNP_LAUNCH_FINISH finalize the cryptographic digest and
stores it as the measurement of the guest at launch.

While finalizing the launch flow, it also issues the LAUNCH_UPDATE
command to encrypt the VMSA pages.

If its an SNP guest, then VMSA was added in the RMP entry as a guest
owned page and also removed from the kernel direct map so flush it later
after it is transitioned back to hypervisor state and restored in the
direct map.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>
Signed-off-by: Harald Hoyer <harald@profian.com>
Signed-off-by: Ashish Kalra <ashish.kalra@amd.com>
[mdr: always measure BSP first to get consistent launch measurements]
Signed-off-by: Michael Roth <michael.roth@amd.com>