x86/sev: handle nested #HV exceptions
We support nested #HV exceptions once the IST stack is
switched out. The HV can always inject an #HV, but as per
GHCB specs, the HV will not inject another #HV, if
PendingEvent.NoFurtherSignal is set and we only clear this
after switching out the IST stack and handling the current #HV.
But there is still a window before the IST stack is
switched out, where a malicious HV can inject nested #HV.
This patch checks the interrupted stack to check if
it is the IST stack, and if so panic as this is
not supported and this nested #HV would have corrupted
the iret frame of the previous #HV on the IST stack.

Signed-off-by: Ashish Kalra <ashish.kalra@amd.com>