x86/sev: Add new HV-Fixed page allocation/free API.
When SEV-SNP is active, the TEE extended command header page and
all output buffers for TEE extended commands (such as used by Seamless
Firmware servicing support) must be in hypervisor-fixed state,
that is assigned to the hypervisor and marked immutable in the RMP entrie(s).
Also if SNP is active, TMPM ring buffer pages must be set to hypervisor-fixed.

Add a new generic in-kernel/SEV interface to allocate/free hypervisor fixed
pages which abstracts hypervisor fixed page allocation/free for device
drivers. The API internally either uses SNP_INIT_EX or
SNP_PAGE_SET_STATE firmware call to transition pages to HV-Fixed page state
depending on whether SNP_INIT has been done or not.

If SNP is not enabled then the allocator is simply a wrapper over alloc_pages(),
free_page().

When the drivers free the pages, they are put back on a free list and future
allocation requests will try to re-use the freed pages from this list. The
freed pages list is garbage collected and the pages are actually freed back
to the page allocator only when X86_SNP_SHUTDOWN is invoked which means
HV-Fixed pages can be reliably freed back to the page allocator.

Signed-off-by: Ashish Kalra <ashish.kalra@amd.com>